requirements.txt
PyMySQL
veryprettytable
prettytable-extras
jinja2
